+++ js/data.js
@@ -1,0 +1,5 @@
+import { getPictures } from './mocks.js';
+
+const pictures = getPictures();
+
+export {pictures};
