+++ js/utils.js
@@ -1,0 +1,7 @@
+function createIdGenerator (prefix = '') {
+  let lastGeneratedId = 0;
+
+  return () => prefix + (lastGeneratedId += 1);
+}
+
+export { createIdGenerator };
